fix : opacity textures using alpha channel like refract
The opacity textures were behaving not as expected :

when i assign a bitmap texture to alpha , the alpha channel is used ( "out_alpha" )
I expected the same behaviour when using the texture for opacity .
But in current main VB  it uses color intensity ("out_intensity")

Example : creating the material for a tree leaf. i have a bitmap with the leaf color and alpha , if i use 'refraction' it works , i can assign the bitmap to "color" and "refract" and get a correct result.
But if i use "opacity" i need another texture with a B/W mask for the alpha ... it'd be easier if could use the alpha channel of the color image, as it works for refraction.
This commit should do that.